+++ src/core/infrastructure/news.infra.ts
@@ -1,8 +1,8 @@
 import { textHttpClient } from "@/shared/lib/http";
 
 /**
- * RSS 피드를 가져오는 함수
+ * 매일경제 사이트맵을 가져오는 함수
  */
-export async function fetchRSSFeed(url: string): Promise<string> {
+export async function fetchMKSitemap(url: string): Promise<string> {
   return textHttpClient.get(url);
 }
